src/api/lambdas/requirements.txt
psycopg2-binary==2.9.9
aws_lambda_powertools